import Tk 402.001 from CPAN
git-cpan-module:   Tk
git-cpan-version:  402.001
git-cpan-authorid: NI-S
git-cpan-file:     authors/id/N/NI/NI-S/Tk402.001.tar.gz

+++ pTk/mTk/generic/tkColor.c
@@ -598,7 +598,7 @@ FindClosestColor(tkwin, desiredColorPtr, actualColorPtr)
 	if (stressPtr->numColors == 0) {
 	    panic("FindClosestColor ran out of colors");
 	}
-	closestDistance = 1e30;
+	closestDistance = (float) 1e30;
 	closest = 0;
 	for (colorPtr = stressPtr->colorPtr, i = 0; i < stressPtr->numColors;
 		colorPtr++, i++) {
